// src/lib/services/models/TransferResponseChitStandardResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { HttpStatusCode } from './HttpStatusCode';
import type { TransferResponse } from './TransferResponse';
export type TransferResponseChitStandardResponse = {
    href?: string | null;
    relations?: Array<string> | null;
    method?: string | null;
    routeName?: string | null;
    routeValues?: any;
    status?: boolean;
    message?: string | null;
    data?: TransferResponse;
    statusCode?: HttpStatusCode;
    errors?: any;
};